feat: Add Github icon component
The code changes add a new component called `Github.svelte` which represents an SVG icon for the Github logo. This component accepts props for size, color, and label. It also exports a `class` prop as `customClass`. The SVG path data is defined within the component.

This commit message follows the established convention of using a prefix to indicate the type of change (`feat` for a new feature). The message is concise and describes the purpose of the code changes.